requirements.txt
streamlit>=1.44.0
streamlit-mic-recorder>=0.0.8
transformers>=4.50.0
torch>=2.2.0
soundfile>=0.12.1
librosa>=0.10.2
numpy>=1.26.0
requests>=2.32.0
python-dotenv>=1.0.1
